requirements.txt
Pillow==7.1.2
pyaes==1.6.1
pyasn1==0.4.8
rsa==4.6
Telethon==1.14.0
